requirements.txt
alembic==1.3.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dlib==19.19.0
Flask==1.1.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Materialize==0.2.4
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
greenlet==0.4.15
idna==2.8
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.1
Mako==1.1.0
MarkupSafe==1.1.1
numpy==1.18.1
opencv-python==4.1.2.30
pickle-mixin==1.0.2
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
requests==2.22.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.12.0
sklearn==0.0
SQLAlchemy==1.3.10
urllib3==1.25.8
Werkzeug==0.16.0
WTForms==2.2.1
